Move vaults transactions to separate folder
