+++ src/transactions/index.ts
@@ -6,3 +6,4 @@ export * from './constants'
 export * as nftLending from './nftLending'
 export * as staking from './staking'
 export * as banxSol from './banxSol'
+export * as vault from './vault'
