+++ src/transactions/nftLending/offers/index.ts
@@ -1,5 +1,3 @@
 export * from './createMakeBondingOfferTxnData'
 export * from './createRemoveOfferTxnData'
 export * from './createUpdateBondingOfferTxnData'
-export * from './createClaimLenderVaultTxnData'
-export * from './createUpdateUserVaultTxnData'
